Bubble web/timeout exceptions to UpdateManager
