+++ src/Squirrel.Client/BitsManager.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net;
 using System.Reactive;
 using System.Reactive.Linq;
 using System.Reactive.Subjects;
@@ -35,16 +34,7 @@ namespace Squirrel.Client
             progress = progress ?? new Subject<int>();
 
             var ret = Http.DownloadUrl(url)
-                .Catch<byte[], TimeoutException>(ex => {
-                    // TODO: log this exception?
-                    return Observable.Return(new byte[0]);
-                })
-                .Catch<byte[], WebException>(ex => {
-                    // TODO: log this exception?
-                    return Observable.Return(new byte[0]);
-                })
-                .Select(x =>
-                {
+                .Select(x => {
                     using (var reader = new StreamReader
                         (new MemoryStream(x), Encoding.UTF8)) {
                         return reader.ReadToEnd();
